requirements.txt
dnspython==2.1.0
email-validator==1.1.3
fhir.resources==6.1.0
jsonlines==2.0.0
pydantic==1.8.2
requests==2.26.0
